+++ tests/helpers/enableCjkForLanguage.js
@@ -1,0 +1,7 @@
+module.exports = function enableCjkForLanguage(lang) {
+    if (!lang) return lang;
+    if (lang.abbreviationUnits || lang.displayAbbreviations) {
+        lang.abbreviationScheme = "cjk";
+    }
+    return lang;
+};